+++ rematch-main/Src/Modules/Infrastructure/CameraProvider/TCPServer.h
@@ -5,14 +5,14 @@
 
 class TCPServer {
 public:
-    TCPServer(int port);
+    TCPServer(uint16_t port);
     ~TCPServer();
     void start();
     void stop();
     void send(const std::vector<unsigned char>& data);
 
 private:
-    int port;
+    uint16_t port;
     int serverFd;
     int clientFd;
     std::thread acceptThread;
